Health Insurance Cost Prediction
You will work with a dataset that includes various inputs such as age, gender, BMI, number of children, smoking habits, and geo-location to predict the output - insurance charges. The dataset contains detailed features like the sex of the insurance contractor, body mass index, number of dependents, smoking status, and the region of residence in the US. The target output to predict is the individual medical costs billed by health insurance.